layout - move restore logic into layout itself
